Session 5
-Fixed previous NOTE issue
-Implemented Save System
-Created Intro Scene
-Added dialogue
-Added text prompt system

#NOTE to future me:
There is currently no connection between clicking a saved file and moving to another screen.

In addition creating a Save File should connect the user to the intro scene after saving it's data.

Also at present clicking a file that has already been saved and entering in a new name overwrites the previous save. Perhaps remove click functionality on save files that are already present and instead add a trash bin icon to the right with a double confirmation prompt to make sure people don't accidentally overwrite their data.

Finally if a user enters a really long name the text goes outside of the box, either add a character limit or a way to resize the text when their name is long (some people got long names #2 probably a better option)